Fix size_t to int conversion warnings (MSVC) in shared_subset.h. Addresses #917
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12041 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/atermpp/include/mcrl2/atermpp/shared_subset.h
@@ -74,7 +74,7 @@ class shared_subset
           : atermpp::aterm_appl(value ? get_true() : get_false())
         {}
 
-        bdd_node(int bit, const bdd_node &true_node, const bdd_node &false_node)
+        bdd_node(size_t bit, const bdd_node &true_node, const bdd_node &false_node)
           : atermpp::aterm_appl(get_node(), atermpp::aterm_int(bit), true_node, false_node)
         {}
 
@@ -93,7 +93,7 @@ class shared_subset
           return function() == get_node();
         }
 
-        int bit()
+        size_t bit()
         {
           return atermpp::aterm_int((*this)[0]).value();
         }
